Move helper functions to common location
