+++ tests/test_commands.py
@@ -1,0 +1,50 @@
+import os
+
+import pytest
+
+from twine.commands import _find_dists, _group_wheel_files_first
+
+
+def test_ensure_wheel_files_uploaded_first():
+    files = _group_wheel_files_first(
+        ["twine/foo.py", "twine/first.whl", "twine/bar.py", "twine/second.whl"]
+    )
+    expected = [
+        "twine/first.whl",
+        "twine/second.whl",
+        "twine/foo.py",
+        "twine/bar.py",
+    ]
+    assert expected == files
+
+
+def test_ensure_if_no_wheel_files():
+    files = _group_wheel_files_first(["twine/foo.py", "twine/bar.py"])
+    expected = ["twine/foo.py", "twine/bar.py"]
+    assert expected == files
+
+
+def test_find_dists_expands_globs():
+    files = sorted(_find_dists(["twine/__*.py"]))
+    expected = [
+        os.path.join("twine", "__init__.py"),
+        os.path.join("twine", "__main__.py"),
+    ]
+    assert expected == files
+
+
+def test_find_dists_errors_on_invalid_globs():
+    with pytest.raises(ValueError):
+        _find_dists(["twine/*.rb"])
+
+
+def test_find_dists_handles_real_files():
+    expected = [
+        "twine/__init__.py",
+        "twine/__main__.py",
+        "twine/cli.py",
+        "twine/utils.py",
+        "twine/wheel.py",
+    ]
+    files = _find_dists(expected)
+    assert expected == files
